Chore: unexport Planner internal methods
