+++ planner/planner.go
@@ -148,7 +148,7 @@ func (planner *MyPlanner) Init(mapsClientApiKey string, redisURL *url.URL, redis
 	}
 }
 
-func (planner *MyPlanner) SingleDayNearbySearchHandler(context *gin.Context) {
+func (planner *MyPlanner) singleDayNearbySearchHandler(context *gin.Context) {
 	country := context.DefaultQuery("country", "USA")
 	city := context.DefaultQuery("city", "San Diego")
 	radius := context.DefaultQuery("radius", "10000")
@@ -187,7 +187,7 @@ func (planner *MyPlanner) Destroy() {
 	planner.RedisClient.Destroy()
 }
 
-func (planner *MyPlanner) ReverseGeocodingHandler(context *gin.Context) {
+func (planner *MyPlanner) reverseGeocodingHandler(context *gin.Context) {
 	latitude, _ := strconv.ParseFloat(context.Query("lat"), 64)
 	longitude, _ := strconv.ParseFloat(context.Query("lng"), 64)
 	result, err := planner.Solver.Searcher.GetMapsClient().ReverseGeocode(context, latitude, longitude)
@@ -235,7 +235,7 @@ func (planner *MyPlanner) removePlacesMigrationHandler(context *gin.Context) {
 	}
 }
 
-func (planner *MyPlanner) PlaceStatsHandler(context *gin.Context) {
+func (planner *MyPlanner) placeStatsHandler(context *gin.Context) {
 	var placeCount int
 	var err error
 	if _, placeCount, err = planner.RedisClient.GetPlaceCountInRedis(context); err != nil {
@@ -266,7 +266,7 @@ type GeocodeCityView struct {
 	Cities map[string]string
 }
 
-func (planner *MyPlanner) GetCitiesHandler(context *gin.Context) {
+func (planner *MyPlanner) getCitiesHandler(context *gin.Context) {
 	views := toCityViews(geocodes)
 	term := context.DefaultQuery("term", "")
 	term = strings.ToLower(term)
@@ -286,7 +286,7 @@ func (planner *MyPlanner) GetCitiesHandler(context *gin.Context) {
 	context.JSON(http.StatusOK, gin.H{"results": results})
 }
 
-func (planner *MyPlanner) CityStatsHandler(context *gin.Context) {
+func (planner *MyPlanner) cityStatsHandler(context *gin.Context) {
 	cityCount := len(geocodes)
 	view := GeocodeCityView{
 		Count:  cityCount,
@@ -314,7 +314,7 @@ func (planner *MyPlanner) Planning(ctx context.Context, planningRequest *Plannin
 		Timestamp: time.Now().Format(time.RFC3339),
 	}
 	planner.PlanningEvents <- event
-	planner.PlanningEventLogging(event)
+	planner.planningEventLogging(event)
 
 	if len(planningResponse.Solutions) == 0 {
 		resp.Err = errors.New("cannot find a valid solution")
@@ -643,7 +643,7 @@ func (planner *MyPlanner) oauthCallback(ctx *gin.Context) {
 	}
 }
 
-func (planner MyPlanner) UserClickOnEmailVerification(ctx *gin.Context) {
+func (planner *MyPlanner) userClickOnEmailVerification(ctx *gin.Context) {
 	code := ctx.DefaultQuery("code", "")
 	if err := planner.RedisClient.CreateUserOnEmailVerified(ctx, code); err != nil {
 		ctx.String(http.StatusBadRequest, "failed to verify user email, please contact Vacation Planner")
@@ -651,7 +651,7 @@ func (planner MyPlanner) UserClickOnEmailVerification(ctx *gin.Context) {
 	ctx.Redirect(http.StatusMovedPermanently, "/v1/log-in")
 }
 
-func (planner MyPlanner) SetupRouter(serverPort string) *http.Server {
+func (planner *MyPlanner) SetupRouter(serverPort string) *http.Server {
 	gin.SetMode(gin.ReleaseMode)
 	if planner.Environment == "debug" {
 		gin.SetMode(gin.DebugMode)
@@ -675,14 +675,14 @@ func (planner MyPlanner) SetupRouter(serverPort string) *http.Server {
 		v1.GET("/", planner.searchPageHandler)
 		v1.GET("/plans", planner.getPlanningApi)
 		v1.POST("/signup", planner.UserEmailVerify)
-		v1.GET("/verify", planner.UserClickOnEmailVerification)
-		v1.POST("/login", planner.UserLogin)
-		v1.GET("/reverse-geocoding", planner.ReverseGeocodingHandler)
-		v1.GET("/single-day-nearby-search", planner.SingleDayNearbySearchHandler)
+		v1.GET("/verify", planner.userClickOnEmailVerification)
+		v1.POST("/login", planner.userLogin)
+		v1.GET("/reverse-geocoding", planner.reverseGeocodingHandler)
+		v1.GET("/single-day-nearby-search", planner.singleDayNearbySearchHandler)
 		v1.GET("/log-in", planner.login)
 		v1.GET("/sign-up", planner.signup)
 		v1.GET("/plans/:id", planner.getPlanDetails)
-		v1.GET("/cities", planner.GetCitiesHandler)
+		v1.GET("/cities", planner.getCitiesHandler)
 		v1.POST("/customize", planner.customize)
 		v1.GET("/template", planner.customizedTemplate)
 		v1.GET("/login-google", planner.handleLogin)
@@ -697,17 +697,17 @@ func (planner MyPlanner) SetupRouter(serverPort string) *http.Server {
 		v1.GET("/profile", planner.profile)
 		users := v1.Group("/users")
 		{
-			users.POST("/:username/plans", planner.UserSavedPlansPostHandler)
-			users.GET("/:username/plans", planner.UserSavedPlansGetHandler)
-			users.DELETE("/:username/plan/:id", planner.UserPlanDeleteHandler)
+			users.POST("/:username/plans", planner.userSavedPlansPostHandler)
+			users.GET("/:username/plans", planner.userSavedPlansGetHandler)
+			users.DELETE("/:username/plan/:id", planner.userPlanDeleteHandler)
 		}
 	}
 
 	// API endpoints for collecting database statistics
 	stats := myRouter.Group("/stats")
 	{
-		stats.GET("places", planner.PlaceStatsHandler)
-		stats.GET("cities", planner.CityStatsHandler)
+		stats.GET("places", planner.placeStatsHandler)
+		stats.GET("cities", planner.cityStatsHandler)
 	}
 
 	svr := &http.Server{
