+++ planner/streaming.go
@@ -10,7 +10,7 @@ import (
 	"golang.org/x/text/language"
 )
 
-func (planner MyPlanner) PlanningEventLogging(event iowrappers.PlanningEvent) {
+func (planner *MyPlanner) planningEventLogging(event iowrappers.PlanningEvent) {
 	eventData := map[string]string{
 		"user":      event.User,
 		"city":      event.City,
@@ -20,7 +20,7 @@ func (planner MyPlanner) PlanningEventLogging(event iowrappers.PlanningEvent) {
 	planner.RedisClient.StreamsLogging(planner.RedisStreamName, eventData)
 }
 
-func (planner MyPlanner) ProcessPlanningEvent(worker int, wg *sync.WaitGroup) {
+func (planner *MyPlanner) ProcessPlanningEvent(worker int, wg *sync.WaitGroup) {
 	for event := range planner.PlanningEvents {
 		c := cases.Title(language.English)
 		log.Debugf("worker %d processing event for %s", worker, c.String(event.City)+", "+strings.ToUpper(event.Country))
